wip: refactor out header components
